Move Admin, Import, License, Settings, and Upgrade to an Admin namespace
